+++ test/run_unit_tests.sql
@@ -1,0 +1,2 @@
+prompt RUN UNIT TESTS
+execute ut.run('test_om_tapigen');